pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "access"
dynamic = ["version"]
dependencies = [
    "Click",
    "argon2",
    "base58",
    "protobuf >= 4.25",
    "pycryptodome",
    "pynacl",
]

[project.scripts]
access = 'uplink.scripts.main:main'

[project.optional-dependencies]
dev = [
    "black >=22.3.0",
    "mypy-protobuf >=3.5.0"
]
test = [
    "pytest >=7",
    "tox",
]

[tool.black]
extend-exclude = '.*_pb2.py'

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
